fix(gpu): fix compression benchmarking
